+++ wasm/CHANGELOG.md
@@ -7,3 +7,10 @@
 ## 0.1.8
 
 - Fixed Group struct to have an optional colour field.
+
+## 0.1.9
+
+- Fixed Rank struct to have an optional country field.
+- Added UserAccountHistory struct and included it in the User struct as an optional field.
+- Added Banner struct and included it in the User struct as optional fields.
+- Added Badge struct and included it in the User struct as an optional field.